Fixed findbugs issues
Configured findbugs for xml output.
And updated the maven release plugin

+++ src/main/java/hudson/plugins/build_timeout/impl/ElasticTimeOutStrategy.java
@@ -95,7 +95,7 @@ public class ElasticTimeOutStrategy extends BuildTimeOutStrategy {
         }
 
 
-        return nonFailingBuilds > 0 ? durationSum / nonFailingBuilds : 0;
+        return nonFailingBuilds > 0 ? ((double)durationSum) / nonFailingBuilds : 0;
     }
 
     public Descriptor<BuildTimeOutStrategy> getDescriptor() {
